Instead of rendering in a loop, first populate vertex array with resulting pixel and draw it to the screen later

+++ colortable.h
@@ -56,11 +56,11 @@ public:
 	using iterator = colormap_t::iterator;
 	using const_iterator = colormap_t::const_iterator;
 
-	ColorTable(int size)
+	ColorTable(unsigned int size)
 	{
 		for (size_t i = 0; i < size; i++)
 		{
-			auto hsl = HSLColor(240 / size * static_cast<double>(i), 1.0, 0.5);
+			auto hsl = HSLColor(240 / static_cast<double>(size) * static_cast<double>(i), 1.0, 0.5);
 			auto rgb = HSLToRGB(hsl);
 			colormap[i] = sf::Color(rgb.R, rgb.G, rgb.B);
 		}
